Add support for more output formats
Add support for all output formats supported by Grype, except for the Go template output format.

Always set the `report` output paramter, for both SARIF and other report formats. This allows that output to be used for all cases and prepares for future deprecation/removal of the `sarif` output parameters.

Signed-off-by: Maarten Hazewinkel <maarten.hazewinkel@gmail.com>